Use `github:` to reference files in GitHub repos
Previously we were using `https://github.com/REPO/FILE`, but that is an
invalid URL so could be confusing. `github:REPO:file` is much clearer
that we are referring to a file within a GitHub repository, without
any expectation of the argument being a real URL you could use in
a user agent.