Fix failed sudo being a successful run
In order to accept the Xcode licence, sudo is needed, so suited checks
before running that sudo can be run. If the check fails (too many
password errors, not allowed to sudo, ^C pressed) suited reported the
run has succeeded. This was incorrect and is now fixed.

As sudo is rarely needed otherwise, there is also a new `-n` flag which
suppresses the checked.