Create the `.ssh` directory if it doesn't exist
This line checks for the `known_hosts` file, but it assumes that the `.ssh` directory itself exists. On a fresh computer, this command failed.

This addition adds two more arguments to the expression to create the directory if it doesn't exist and give it the correct permissions. They fail silently so if they do already exist it will still have the desired effect.