Create the `.ssh` directory if it doesn't exist
On a brand new machine, suited will try to pre-populate the ssh
`known_hosts` file with the fingerprints of `github.com`. This
stops it prompting "y/n" on the first connection made -- which
could be some time after running suited, and you may not notice
for a while, interrupting the setup.

As I always copy my SSH key by hand to a new computer before running
suited, I hadn't spotted that this would fail if the `.ssh`
directory wasn't there.

Thanks to @annashipman for the original PR #7 that prompted this.