Stop using suited subshells
Previously, to set a new root context--so that relative filenames
worked--suited would re-execute itself with the new suitfile. This
would cause error reporting to be hidden away, and multiple "hey it
worked" messages would appear, which was confusing.

Now we just call `process_suitfile` again, and use the `BASE` variable
instead. Recursion!