Assume `suitfile` in directories, or fall back
When naming just a directory in a suitfile (eg 'setup/antirsi/'),
suited would look for Brewfile, bootstrap script, Gemfile and
crontab files and use them if they existed.

A better option is to assume a `suitfile` exists and use that,
so you have the full range of commands suited can perform. For
backwards compatibility, the previous behaviour is preserved if
no `suitfile` exists.