Recommend using the latest version, not master
As master can contain breaking changes to suited's behaviour, the
instructions shouldn't recommend fetching it. Instead, I've created
a `latest` branch to point at the commit of the most recent proper
release of suited.

As I'm bound to forget to update this at some point in the future,
I've also created a release checklist to use in future.