Clarify how to make a release in the checklist
It should be done via Github, so that the changelog text can be used
as the description. Although you can tag the version in git and push
it, you still need to go to Github to paste in the description, so
you might as well create the release there in the first place.