Add `needenv` command to ensure an envvar is set
Some setup scripts require environment variables to work correctly. With
the `needenv` command, `suited` can abort if a required environment
variable is not set to a value.

A script can already exit with an error if the variable is unset,
and this will stop suited at that point. But the script could treat the
variable as optional whilst your setup does not. And it is another
way of making requirements explicit when looking at setup code, without
having to delve into the script as well.